Move eventLink from case folder to the event one #3133
Because this component no longer is only for cases, and will be made
more generic (It will at least be used for contacts)
