+++ sormas-api/src/main/resources/captions_de-CH.properties
@@ -910,7 +910,7 @@ eventEventsList=Ereignisliste
 eventNewEvent=Neues Ereignis
 eventSearchEvent=Ereignis suchen
 eventSearchSpecificEvent=Nach einem bestimmten Ereignis suchen
-linkEventToCase=Ereignis verknüpfen
+linkEvent=Ereignis verknüpfen
 eventSelect=Ereignis auswählen 
 eventDefaultView=Ereignisse
 eventActionsView=Aktionen
