+++ sormas-api/src/main/resources/captions_de-DE.properties
@@ -910,7 +910,7 @@ eventEventsList=Ereignisliste
 eventNewEvent=Neues Ereignis
 eventSearchEvent=Ereignis suchen
 eventSearchSpecificEvent=Nach einem bestimmten Ereignis suchen
-linkEventToCase=Ereignis verknüpfen
+linkEvent=Ereignis verknüpfen
 eventSelect=Ereignis wählen 
 eventDefaultView=Ereignisse
 eventActionsView=Aktionen
