+++ sormas-api/src/main/resources/captions_es-CU.properties
@@ -910,7 +910,7 @@ eventEventsList=Lista de eventos
 eventNewEvent=Nuevo evento
 eventSearchEvent=Buscar evento
 eventSearchSpecificEvent=Buscar evento específico
-linkEventToCase=Vincular evento
+linkEvent=Vincular evento
 eventSelect=Seleccionar evento 
 eventDefaultView=Eventos
 eventActionsView=Acciones
