+++ sormas-api/src/main/resources/captions_es-EC.properties
@@ -910,7 +910,7 @@ eventEventsList=Lista de eventos
 eventNewEvent=Nuevo evento
 eventSearchEvent=Search event
 eventSearchSpecificEvent=Search specific Event
-linkEventToCase=Link event
+linkEvent=Link event
 eventSelect=Select event 
 eventDefaultView=Events
 eventActionsView=Actions
