+++ sormas-api/src/main/resources/captions_fi-FI.properties
@@ -910,7 +910,7 @@ eventEventsList=Tapahtumalista
 eventNewEvent=Uusi tapahtuma
 eventSearchEvent=Hae tapahtumaa
 eventSearchSpecificEvent=Etsi tietty tapahtuma
-linkEventToCase=Link event
+linkEvent=Link event
 eventSelect=Select event 
 eventDefaultView=Events
 eventActionsView=Actions
