+++ sormas-api/src/main/resources/captions_fr-CH.properties
@@ -910,7 +910,7 @@ eventEventsList=Liste d'événements
 eventNewEvent=Nouvel événement
 eventSearchEvent=Rechercher un événement
 eventSearchSpecificEvent=Rechercher un événement spécifique
-linkEventToCase=Lier un événement
+linkEvent=Lier un événement
 eventSelect=Sélectionnez un événement 
 eventDefaultView=Événements
 eventActionsView=Actions
