+++ sormas-api/src/main/resources/captions_it-CH.properties
@@ -910,7 +910,7 @@ eventEventsList=Lista Eventi
 eventNewEvent=Nuovo evento
 eventSearchEvent=Cerca evento
 eventSearchSpecificEvent=Cerca evento specifico
-linkEventToCase=Collega evento
+linkEvent=Collega evento
 eventSelect=Seleziona evento 
 eventDefaultView=Eventi
 eventActionsView=Azioni
