+++ sormas-api/src/main/resources/captions_it-IT.properties
@@ -910,7 +910,7 @@ eventEventsList=Lista Eventi
 eventNewEvent=Nuovo evento
 eventSearchEvent=Cerca evento
 eventSearchSpecificEvent=Cerca evento specifico
-linkEventToCase=Link event
+linkEvent=Link event
 eventSelect=Select event 
 eventDefaultView=Events
 eventActionsView=Actions
