+++ sormas-api/src/main/resources/captions_ku-TR.properties
@@ -910,7 +910,7 @@ eventEventsList=Events list
 eventNewEvent=New event
 eventSearchEvent=Search event
 eventSearchSpecificEvent=Search specific Event
-linkEventToCase=Link event
+linkEvent=Link event
 eventSelect=Select event 
 eventDefaultView=Events
 eventActionsView=Actions
