+++ sormas-ui/src/main/java/de/symeda/sormas/ui/events/eventLink/EventList.java
@@ -18,7 +18,7 @@
  * ******************************************************************************
  */
 
-package de.symeda.sormas.ui.caze.eventLink;
+package de.symeda.sormas.ui.events.eventLink;
 
 import java.util.List;
